ImageClassifier/Aryan Mishra Live Video Frame Classifier/requirements.txt
streamlit
opencv-python
Pillow
numpy
tensorflow
streamlit-webrtc
av